Math Update
Was able to (sort of ) store the students with more than one successes or more than one failures! Just need to figure out a way to organize the data so a student's ID doesn't appear more than once. (basically a way to append new course to existing array with that student's ID]